+++ FileTransferTool/Models/Chunk.cs
@@ -5,7 +5,7 @@
     /// </summary>
     public class Chunk
     {
-        public int ChunkId { get; set; }
+        public long ChunkId { get; set; }
         public long Offset { get; set; }
         public int Size { get; set; }
         public byte[] Buffer { get; set; }
